+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # 2.x branch
 ## 2.3 branch
+### 2.3.3
+* improved the code to check the cache validity. Removed `checkIfCacheIsValid()` and `getNextToBePublished()` methods;
+* updated Facebook's tags.
+
+### 2.3.2
+* logged users can view future videos and drafts;
+
 ### 2.3.1
 * added userbar for frontend. It allows to edit an delete videos.